border: move extraction code to new file
Extraction stage don't rely on gcc plugin to get code location now,
it's better to move it to a seperate file. After moving, the code
framework becomes clear, three seperate files correspond to the
three stages of preprocessing.
