+++ sched_boundary/extract.py
@@ -1,0 +1,172 @@
+#!/usr/bin/env python3
+# Copyright 2019-2022 Alibaba Group Holding Limited.
+# SPDX-License-Identifier: GPL-2.0 OR BSD-3-Clause
+
+from collections import defaultdict
+from itertools import groupby as _groupby
+from yaml import load, dump, resolver, CLoader as Loader, CDumper as Dumper
+import json
+import re
+import os
+import sys
+
+# Use set as the default sequencer for yaml
+Loader.add_constructor(resolver.BaseResolver.DEFAULT_SEQUENCE_TAG,
+                     lambda loader, node: set(loader.construct_sequence(node)))
+Dumper.add_representer(set, lambda dumper, node: dumper.represent_list(node))
+
+# tmp directory to store middle files
+tmpdir = None
+
+# directory to store schedule module source code
+modpath = None
+
+class SchedBoundaryExtract(object):
+    def __init__(self, src_file, tmpdir, modpath):
+        with open(tmpdir + 'sched_boundary_extract.yaml') as f:
+            self.config = load(f, Loader)
+
+        self.src_file = src_file
+        self.modpath = modpath
+        self.mod_files = self.config['mod_files']
+        self.mod_srcs = {f for f in self.mod_files if f.endswith('.c')}
+        self.mod_hdrs = self.mod_files - self.mod_srcs
+        self.fn_list = []
+        self.fn_ptr_list = []
+        self.interface_list = []
+        self.var_list = []
+
+        if src_file in self.mod_hdrs:
+            file_name = tmpdir + 'header_symbol.json'
+        else:
+            file_name = src_file + '.sched_boundary'
+
+        with open(file_name) as f:
+            metas = json.load(f)
+            self.meta_fn = metas['fn']
+            self.meta_var = metas['var']
+
+    def function_location(self):
+        unique = set()
+        for fn in self.meta_fn:
+            # filter out *.h in *.c
+            if fn['file'] != self.src_file: continue
+
+            # remove duplicated function
+            obj = tuple(fn['signature'])
+            if obj in unique: continue
+            unique.add(obj)
+
+            if obj in self.config['function']['sched_outsider'] or \
+               obj in self.config['function']['init']:
+                self.fn_list.append(fn),
+            elif obj in self.config['function']['fn_ptr']:
+                self.fn_ptr_list.append(fn)
+            elif obj in self.config['function']['interface']:
+                self.interface_list.append(fn)
+
+    def var_location(self):
+        for var in self.meta_var:
+            if var['file'] != self.src_file:
+                continue
+            if var['name'] in self.config['global_var']['force_private']:
+                continue
+            # share public (non-static) variables by default
+            if var['public'] or var['name'] in self.config['global_var']['extra_public']:
+                self.var_list.append(var)
+
+    def function_extract(self, lines, fn_export_jump):
+        for fn in self.fn_list:
+            name, inline = fn['name'], fn['inline']
+            (row_start,col_start), (row_end,_) = fn['l_brace_loc'], fn['r_brace_loc']
+
+            if inline or \
+                    tuple(fn['signature']) in self.config['function']['optimized_out']:
+                lines[row_end] += "/* DON'T MODIFY FUNCTION {}, IT'S NOT PART OF SCHEDMOD */\n".format(name)
+            else:
+                # convert function body "{}" to ";"
+                # only handle normal kernel function definition
+                lines[row_start] = lines[row_start][:col_start] + ";\n"
+                for i in range(row_start+1, row_end+1):
+                    lines[i] = ''
+
+        fn_ptr_export_header_fmt = "PLUGSCHED_FN_PTR({fn}, {ret}, {params})\n"
+        fn_ptr_export_c_fmt = "extern {ret} {fn}({params});\n"
+        for fn in self.fn_ptr_list:
+            name, decl_str = fn['name'], fn['decl_str']
+            (row_start,col_start), (row_end,_) = fn['name_loc'], fn['r_brace_loc']
+
+            fn_export_jump.write(fn_ptr_export_header_fmt.format(**decl_str))
+            new_name = '__mod_' + name
+            lines[row_start] = lines[row_start][:col_start] + \
+                lines[row_start][col_start:].replace(name, '__used ' + new_name)
+            lines[row_end] = lines[row_end] + '\n' + \
+                "/* DON'T MODIFY SIGNATURE OF FUNCTION {}, IT'S CALLBACK FUNCTION */\n".format(new_name) + \
+                fn_ptr_export_c_fmt.format(**decl_str)
+
+        interface_export_fmt = "EXPORT_PLUGSCHED({fn}, {ret}, {params})\n"
+        for fn in self.interface_list:
+            name, decl_str, (row_end,_) = fn['name'], fn['decl_str'], fn['r_brace_loc']
+            fn_export_jump.write(interface_export_fmt.format(**decl_str))
+
+            # everyone know that syscall ABI should be consistent
+            if any(name.startswith(prefix) for prefix in self.config['interface_prefix']):
+                continue
+            lines[row_end] += \
+                "/* DON'T MODIFY SIGNATURE OF FUNCTION {}, IT'S INTERFACE FUNCTION */\n".format(name)
+
+    def var_extract(self, lines):
+        # General handling all shared variables
+        orig_lines = list(lines)
+        for var in list(self.var_list):
+            name, row_start, row_end = var['name'], var['decl_start_line'], var['decl_end_line']
+
+            # delete data initialization code
+            for i in range(row_start+1, row_end+1):
+                lines[i] = ''
+
+            # Specially handling shared per_cpu and static_key variables to improve readability
+            line = orig_lines[row_start]
+            if 'DEFINE_PER_CPU(' in line:
+                line = line.replace('DEFINE_PER_CPU(', 'DECLARE_PER_CPU(').replace('static ', '')
+            elif 'DEFINE_PER_CPU_SHARED_ALIGNED(' in line:
+                line = line.replace('DEFINE_PER_CPU_SHARED_ALIGNED(', 'DECLARE_PER_CPU_SHARED_ALIGNED(').replace('static ', '')
+            elif 'DEFINE_STATIC_KEY_FALSE(' in line:
+                line = line.replace('DEFINE_STATIC_KEY_FALSE(', 'DECLARE_STATIC_KEY_FALSE(').replace('static ', '')
+            elif 'DEFINE_STATIC_KEY_TRUE(' in line:
+                line = line.replace('DEFINE_STATIC_KEY_TRUE(', 'DECLARE_STATIC_KEY_TRUE(').replace('static ', '')
+            elif 'EXPORT_' in line and '_SYMBOL' in line:
+                line = ''
+            else:
+                # delete data definition
+                lines[row_start] = ''
+                continue
+
+            lines[row_start] = line
+            self.var_list.remove(var)
+
+        # convert data definition to declarition
+        for var in self.var_list:
+            row_start = var['decl_start_line']
+            lines[row_start] += var['decl_str'] + '\n'
+
+    def extract_file(self):
+        src_f = self.src_file
+        self.function_location()
+        self.var_location()
+
+        with open(src_f) as in_f, open(src_f + '.export_jump.h', 'w') as fn_export_jump:
+            lines = in_f.readlines()
+            self.function_extract(lines, fn_export_jump)
+            self.var_extract(lines)
+
+            with open(self.modpath + os.path.basename(src_f), 'w') as out_f:
+                out_f.writelines(lines)
+
+if __name__ == '__main__':
+
+    src_file = sys.argv[1]
+    tmpdir = sys.argv[2]
+    modpath = sys.argv[3]
+    extract = SchedBoundaryExtract(src_file, tmpdir, modpath)
+    extract.extract_file()
