+++ sched_boundary/sched_boundary.py
@@ -20,28 +20,6 @@ tmpdir = None
 # directory to store schedule module source code
 modpath = None
 
-class SchedBoundary(object):
-    def __init__(self, config):
-        with open(config) as f:
-            self.config = load(f, Loader)
-        self.mod_files = self.config['mod_files']
-        self.mod_srcs = {f for f in self.mod_files if f.endswith('.c')}
-        self.mod_hdrs = self.mod_files - self.mod_srcs
-
-    def process_passes(self, p, _):
-        if p.name != '*free_lang_data':
-            return
-        self.final_work()
-
-    def register_cbs(self):
-        if hasattr(self, 'var_declare'):
-            gcc.register_callback(gcc.PLUGIN_FINISH_DECL, self.var_declare)
-        if hasattr(self, 'final_work'):
-            gcc.register_callback(gcc.PLUGIN_PASS_EXECUTION, self.process_passes)
-        if hasattr(self, 'include_file'):
-            gcc.register_callback(gcc.PLUGIN_INCLUDE_FILE, self.include_file)
-
-# Workarounds
 class GccBugs(object):
     array_pointer_re = re.compile(r'(.*)\[([0-9]*)\] \*\s*([^,\);]*)')
 
@@ -102,153 +80,14 @@ class GccBugs(object):
         if decl.str_decl.find("...") >= 0:
             signature["params"] += ", ..."
 
-class SchedBoundaryExtract(SchedBoundary):
+
+class SchedBoundaryCollect(object):
     def __init__(self):
-        super().__init__(tmpdir + 'sched_boundary_extract.yaml')
-        self.fn_dict = {}
-        self.fn_ptr_list = []
-        self.interface_list = []
-        self.var_list = []
-        self.fake = 'kernel/sched/fake.c'
-
-    def register_cbs(self):
-        if gcc.get_main_input_filename() in self.mod_srcs | {self.fake}:
-            super().register_cbs()
-
-    def function_location(self):
-        src_file = gcc.get_main_input_filename()
-        tmp_config = src_file + ".sched_boundary"
-        self.fn_dict.setdefault(src_file, list())
-
-        # FIXME: there is not fake.c.sched_boundary
-        if src_file == self.fake: return
-
-        with open(tmp_config) as f:
-            self.meta = json.load(f)
-
-        for fn in self.meta['fn']:
-            obj = tuple(fn['signature'])
-            if obj in self.config['function']['sched_outsider'] or \
-               obj in self.config['function']['init']:
-                self.fn_dict[src_file].append(fn),
-            elif obj in self.config['function']['fn_ptr']:
-                self.fn_ptr_list.append(fn)
-            elif obj in self.config['function']['interface']:
-                self.interface_list.append(fn)
-
-    def var_location(self):
-        loc = gcc.get_location()
-        src_file = gcc.get_main_input_filename()
-        tmp_config = src_file + ".sched_boundary"
-
-        # FIXME: there is not fake.c.sched_boundary
-        if src_file == self.fake: return
-
-        with open(tmp_config) as f:
-            self.meta = json.load(f)
-
-        for var in self.meta['var']:
-            if var['name'] in self.config['global_var']['force_private']:
-                continue
-            # share public (non-static) variables by default
-            if var['public'] or var['name'] in self.config['global_var']['extra_public']:
-                self.var_list.append(var)
-
-    def final_work(self):
-        self.function_location()
-        self.var_location()
-
-        if gcc.get_main_input_filename() == self.fake:
-            # no function definition in sched-pelt.h
-            for header in self.mod_hdrs:
-                self.fn_dict.setdefault(header, list())
-
-        for src_f in self.fn_dict.keys():
-            self.extract_file(src_f)
-
-    def extract_file(self, src_f):
-        with open(src_f) as in_f, open(src_f + '.export_jump.h', 'w') as fn_export_jump:
-            lines = in_f.readlines()
-
-            for fn in self.fn_dict[src_f]:
-                name, inline = fn['name'], fn['inline']
-                (row_start,col_start), (row_end,_) = fn['l_brace_loc'], fn['r_brace_loc']
-
-                if inline or \
-                        (name, src_f) in self.config['function']['optimized_out']:
-                    lines[row_end] += "/* DON'T MODIFY FUNCTION {}, IT'S NOT PART OF SCHEDMOD */\n".format(name)
-                else:
-                    # convert function body "{}" to ";"
-                    # only handle normal kernel function definition
-                    lines[row_start] = lines[row_start][:col_start] + ";\n"
-                    for i in range(row_start+1, row_end+1):
-                        lines[i] = ''
-
-            fn_ptr_export_header_fmt = "PLUGSCHED_FN_PTR({fn}, {ret}, {params})\n"
-            fn_ptr_export_c_fmt = "extern {ret} {fn}({params});\n"
-            for fn in self.fn_ptr_list:
-                name, decl_str = fn['name'], fn['decl_str']
-                (row_start,col_start), (row_end,_) = fn['name_loc'], fn['r_brace_loc']
-
-                fn_export_jump.write(fn_ptr_export_header_fmt.format(**decl_str))
-                new_name = '__mod_' + name
-                lines[row_start] = lines[row_start][:col_start] + \
-                    lines[row_start][col_start:].replace(name, '__used ' + new_name)
-                lines[row_end] = lines[row_end] + '\n' + \
-                    "/* DON'T MODIFY SIGNATURE OF FUNCTION {}, IT'S CALLBACK FUNCTION */\n".format(new_name) + \
-                    fn_ptr_export_c_fmt.format(**decl_str)
-
-            interface_export_fmt = "EXPORT_PLUGSCHED({fn}, {ret}, {params})\n"
-            for fn in self.interface_list:
-                name, decl_str, (row_end,_) = fn['name'], fn['decl_str'], fn['r_brace_loc']
-                fn_export_jump.write(interface_export_fmt.format(**decl_str))
-
-                # everyone know that syscall ABI should be consistent
-                if any(name.startswith(prefix) for prefix in self.config['interface_prefix']):
-                    continue
-                lines[row_end] += \
-                    "/* DON'T MODIFY SIGNATURE OF FUNCTION {}, IT'S INTERFACE FUNCTION */\n".format(name)
-
-            # General handling all shared variables
-            orig_lines = list(lines)
-            for var in list(self.var_list):
-                name, row_start, row_end = var['name'], var['decl_start_line'], var['decl_end_line']
-
-                # delete data initialization code
-                for i in range(row_start+1, row_end+1):
-                    lines[i] = ''
-
-                # Specially handling shared per_cpu and static_key variables to improve readability
-                line = orig_lines[row_start]
-                if 'DEFINE_PER_CPU(' in line:
-                    line = line.replace('DEFINE_PER_CPU(', 'DECLARE_PER_CPU(').replace('static ', '')
-                elif 'DEFINE_PER_CPU_SHARED_ALIGNED(' in line:
-                    line = line.replace('DEFINE_PER_CPU_SHARED_ALIGNED(', 'DECLARE_PER_CPU_SHARED_ALIGNED(').replace('static ', '')
-                elif 'DEFINE_STATIC_KEY_FALSE(' in line:
-                    line = line.replace('DEFINE_STATIC_KEY_FALSE(', 'DECLARE_STATIC_KEY_FALSE(').replace('static ', '')
-                elif 'DEFINE_STATIC_KEY_TRUE(' in line:
-                    line = line.replace('DEFINE_STATIC_KEY_TRUE(', 'DECLARE_STATIC_KEY_TRUE(').replace('static ', '')
-                elif 'EXPORT_' in line and '_SYMBOL' in line:
-                    line = ''
-                else:
-                    # delete data definition
-                    lines[row_start] = ''
-                    continue
-
-                lines[row_start] = line
-                self.var_list.remove(var)
-
-            # convert data definition to declarition
-            for var in self.var_list:
-                row_start = var['decl_start_line']
-                lines[row_start] += var['decl_str'] + '\n'
-
-            with open(modpath + os.path.basename(src_f), 'w') as out_f:
-                out_f.writelines(lines)
-
-class SchedBoundaryCollect(SchedBoundary):
-    def __init__(self):
-        super().__init__(tmpdir + 'sched_boundary.yaml')
+        with open(tmpdir + 'sched_boundary.yaml') as f:
+            self.config = load(f, Loader)
+        self.mod_files = self.config['mod_files']
+        self.mod_srcs = {f for f in self.mod_files if f.endswith('.c')}
+        self.mod_hdrs = self.mod_files - self.mod_srcs
         self.fn_properties = []
         self.var_properties = []
         self.edge_properties = []
@@ -256,6 +95,19 @@ class SchedBoundaryCollect(SchedBoundary):
         self.fn_ptr_properties = []
         self.interface_properties = []
         self.seek_public_field = False
+
+    def process_passes(self, p, _):
+        if p.name != '*free_lang_data':
+            return
+        self.final_work()
+
+    def register_cbs(self):
+        if hasattr(self, 'var_declare'):
+            gcc.register_callback(gcc.PLUGIN_FINISH_DECL, self.var_declare)
+        if hasattr(self, 'final_work'):
+            gcc.register_callback(gcc.PLUGIN_PASS_EXECUTION, self.process_passes)
+        if hasattr(self, 'include_file'):
+            gcc.register_callback(gcc.PLUGIN_INCLUDE_FILE, self.include_file)
 
     def decl_in_section(self, decl, section):
         for name, val in decl.attributes.items():
@@ -473,6 +325,7 @@ class SchedBoundaryCollect(SchedBoundary):
         with open(gcc.get_main_input_filename() + '.sched_boundary', 'w') as f:
             json.dump(collect, f, indent=4)
 
+
 if __name__ == '__main__':
     import gcc
 
@@ -480,11 +333,5 @@ if __name__ == '__main__':
     tmpdir = gcc.argument_dict['tmpdir']
     modpath = gcc.argument_dict['modpath']
 
-    if stage == 'extract':
-        sched_boundary = SchedBoundaryExtract()
-    elif stage == "collect":
-        sched_boundary = SchedBoundaryCollect()
-    else:
-        raise Exception("")
-
+    sched_boundary = SchedBoundaryCollect()
     sched_boundary.register_cbs()
